backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Planting Optimisation Tool Backend Infrastructure"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.122.0",
    "geoalchemy2>=0.18.1",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.5",
    "sqlalchemy>=2.0.44",
    "sqlalchemy-schemadisplay>=2.0",
    "sqlalchemy-utils>=0.42.1",
    "datascience",
    "gis",
    "pyjwt>=2.10.1",
    "psycopg>=3.3.2",
    "psycopg-binary>=3.3.2",
    "psutil>=7.1.3",
    "paracelsus>=0.14.0",
    "mermaid-cli>=0.1.2",
    "httpx>=0.28.1",
    "pre-commit>=4.5.1",
    "earthengine-api>=1.7.4, !=1.7.9",
    "google-auth>=2.47.0",
    "bcrypt==4.0.1",
]

[tool.pytest.ini_options]
# Add the 'backend/' directory to the Python path so tests can find src/ etc.
pythonpath = [
    ".",
    "../gis"
]

testpaths = [
    "tests"
]

asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"

[tool.ruff]
# Ignore unused imports (F401) in all SQLAlchemy model files.
lint.per-file-ignores = { "src/models/*.py" = ["F401"] }
exclude = [
    "alembic/versions/*",
]

[tool.uv]
workspace = { members = ["backend", "datascience", "gis"] }

[tool.uv.sources]
datascience = { path = "../datascience", editable = true }
gis = { path = "../gis", editable = true }
